+++ coderedcms/templates/coderedcms/includes/map_pin_description.html
@@ -1,0 +1,3 @@
+<b>{{ name }}</b>
+<p>{{ address }}</p>
+<p><a href='{{ url }}'>View Location</a></p>
